+++ pwa/sw.js
@@ -86,6 +86,7 @@ const SHELL = [
   './js/personalSpace/v2/projectEngine.js',
   './js/personalSpace/v2/questEngine.js',
   './js/personalSpace/v2/momentum.js',
+  './js/personalSpace/v2/revealSelectors.js',
   './js/personalSpace/v2/companionEngine.js',
   './js/personalSpace/v2/viewModels.js',
   './js/personalSpace/v2/telemetry.js',